Moving functionality from CreateObject2 to CreateObject
