+++ UI_Engine/Query/SearchItems.cs
@@ -116,6 +116,14 @@ namespace BH.Engine.UI
 
         /***************************************************/
 
+        public static IEnumerable<Type> ConstructableTypeItems()
+        {
+            return Engine.Reflection.Query.BHoMTypeList()
+                .Where(x => !x.IsNotImplemented() && !x.IsDeprecated() && x?.GetInterface("IImmutable") == null);
+        }
+
+        /***************************************************/
+
         public static IEnumerable<Type> EnumItems()
         {
             return Engine.Reflection.Query.BHoMEnumList()
